pyproject.toml
[tool.poetry]
name = "pptx-generator"
version = "0.1.0"
description = ""
authors = ["fahadsalamatrandhawa <dreamtwister61814@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-pptx = "^1.0.2"
comtypes = "^1.4.6"
streamlit = "^1.37.1"
pymupdf = "^1.24.9"
pypandoc = "^1.13"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
